refactor: rename githooks package
Resolves #19
