+++ src/main/java/org/dev/spanciuc/maven/mojo/git/hooks/GitHookFileGenerator.java
@@ -1,0 +1,25 @@
+package org.dev.spanciuc.maven.mojo.git.hooks;
+
+import java.io.File;
+
+/**
+ * Provides the necessary methods for git hooks file generation.
+ */
+public interface GitHookFileGenerator {
+
+    /**
+     * Creates a git hook file in directory.
+     *
+     * @param parentDirectory the target directory.
+     * @return generated file.
+     */
+    File generate(File parentDirectory);
+
+    /**
+     * Gets the name of the file.
+     *
+     * @return filename.
+     */
+    String getFileName();
+
+}
