+++ src/test/it/add-git-hooks-aggregate-project/setup.groovy
@@ -1,3 +1,3 @@
-import org.dev.spanciuc.maven.mojo.githooks.it.utils.GitTestUtils
+import org.dev.spanciuc.maven.mojo.git.hooks.it.utils.GitTestUtils
 
 println GitTestUtils.initRepository(basedir, ".git")
