+++ src/test/it/add-git-hooks-single-project/verify.groovy
@@ -1,3 +1,3 @@
-import org.dev.spanciuc.maven.mojo.githooks.it.utils.CommitMsgHookTestUtils
+import org.dev.spanciuc.maven.mojo.git.hooks.it.utils.CommitMsgHookTestUtils
 
 println CommitMsgHookTestUtils.checkCommitMsgHookFile(basedir, ".git/hooks", "commit-msg")
